requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
idna==3.2
pyTelegramBotAPI==4.1.0
python-decouple==3.5
requests==2.26.0
urllib3==1.26.7
